requirements.txt
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
h11==0.14.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.4
outcome==1.3.0.post0
pillow==10.3.0
praw==7.7.1
prawcore==2.4.0
proglog==0.1.10
pycparser==2.22
pydub==0.25.1
PySocks==1.7.1
requests==2.32.3
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.8.0
wsproto==1.2.0